mandatory seems perfect

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: angrodri <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/02 14:30:12 by angrodri          #+#    #+#             */
-/*   Updated: 2022/11/04 21:16:59 by angrodri         ###   ########.fr       */
+/*   Updated: 2022/11/11 14:05:06 by angrodri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,14 +18,14 @@
 # include <stdarg.h>
 # include <stdio.h>
 
-int		ft_putnbr_base(unsigned int nbr, char *base);
+int		ft_putnbr_base(unsigned long int nbr, char *base);
 int		ft_putchar(char c);
 int		ft_putnbr(int n);
 int		ft_putstr(char *s);
 int		ft_atoi(const char *arg);
 int		ft_printf(char const *str, ...);
 char	*ft_itoa(int n);
-int		ft_putptr(unsigned int ptr);
+int		ft_putptr(unsigned long int ptr);
 int		ft_uint_print(unsigned int nbr);
 
 #endif
